+++ packages/rax-waterfall/src/common/index.js
@@ -3,7 +3,7 @@ import View from 'rax-view';
 import ScrollView from 'rax-scrollview';
 import RefreshControl from 'rax-refreshcontrol';
 import cloneElement from 'rax-clone-element';
-
+import omit from 'omit.js';
 class Header extends PureComponent {
   render() {
     return <View {...this.props} />;
@@ -116,7 +116,8 @@ class Waterfall extends Component {
     }));
 
     styles.waterfallColumn.width = columnWidth;
-    return (<ScrollView {...props} ref={this.scrollview}>
+    const omittedProps = omit(props, ['dataSource', 'renderItem', 'renderHeader', 'renderFooter', 'columnWidth', 'columnCount', 'columnGap', 'cellProps', 'leftGap', 'rightGap']);
+    return (<ScrollView {...omittedProps} ref={this.scrollview}>
       {cells}
     </ScrollView>);
   }